Conflicts:
	FallRiver/FallRiver/GamePlayState.cpp

+++ FallRiver/FallRiver/Particle.cpp
@@ -36,19 +36,19 @@ void Particle::Render()
 	if(imageID != -1)
 	{
 		RECT tmpRect = {long(pos.x - scaleX * .5f), long(pos.y - scaleY * .5f), long(pos.x + scaleX * .5f), long(pos.y + scaleY * .5f)};
-		tmpRect.left -= GamePlayState::GetInstance()->GetCamera().x;
-		tmpRect.right -= GamePlayState::GetInstance()->GetCamera().x;
-		tmpRect.top -= GamePlayState::GetInstance()->GetCamera().y;
-		tmpRect.bottom -= GamePlayState::GetInstance()->GetCamera().y;
+		tmpRect.left -= (long)GamePlayState::GetInstance()->GetCamera().x;
+		tmpRect.right -= (long)GamePlayState::GetInstance()->GetCamera().x;
+		tmpRect.top -= (long)GamePlayState::GetInstance()->GetCamera().y;
+		tmpRect.bottom -= (long)GamePlayState::GetInstance()->GetCamera().y;
 		view->DrawStaticTexture(imageID, pos.x, pos.y, scaleX, scaleY, nullptr,0.0f, 0.0f, rot, color);
 	}
 	else
 	{
 		RECT tmpRect = {long(pos.x - scaleX * .5f), long(pos.y - scaleY * .5f), long(pos.x + scaleX * .5f), long(pos.y + scaleY * .5f)};
-		tmpRect.left -= GamePlayState::GetInstance()->GetCamera().x;
-		tmpRect.right -= GamePlayState::GetInstance()->GetCamera().x;
-		tmpRect.top -= GamePlayState::GetInstance()->GetCamera().y;
-		tmpRect.bottom -= GamePlayState::GetInstance()->GetCamera().y;
+		tmpRect.left -= (long)GamePlayState::GetInstance()->GetCamera().x;
+		tmpRect.right -= (long)GamePlayState::GetInstance()->GetCamera().x;
+		tmpRect.top -= (long)GamePlayState::GetInstance()->GetCamera().y;
+		tmpRect.bottom -= (long)GamePlayState::GetInstance()->GetCamera().y;
 		view->DrawRect( tmpRect, (color << 8) >> 24, (color << 16) >> 24, (color << 24) >> 24, color >> 24);
 
 	}
